move users/one to actually configUsers/one
