+++ src/boox/models/config_users.py
@@ -1,0 +1,5 @@
+from boox.models.base import BaseSyncResponse
+
+
+class SyncTokenResponse(BaseSyncResponse[None]):
+    """A response, wtih information aobut token expiry date."""
