+++ tests/models/test_users_models.py
@@ -10,7 +10,6 @@ from boox.models.users import (
     FetchTokenRequest,
     SendVerifyCodeRequest,
     SyncSessionTokenResponse,
-    SyncTokenResponse,
     soft_validate_email,
 )
 
@@ -83,18 +82,6 @@ def test_token_is_public_in_json_dump():
     data = DataToken.model_validate({"token": value})
     dumped = data.model_dump_json()
     assert re.compile(r'{"token":"xyz123"}').fullmatch(dumped)
-
-
-def test_sync_token_response_parses_token_expiration_date_correctly():
-    token_expiry = datetime.datetime.now(tz=datetime.UTC).replace(microsecond=0) + datetime.timedelta(days=180)
-    data = SyncTokenResponse.model_validate({
-        "data": None,
-        "message": "SUCCESS",
-        "result_code": 0,
-        "tokenExpiredAt": int(token_expiry.timestamp()),
-    })
-    assert not data.data
-    assert data.token_expired_at == token_expiry
 
 
 def test_sync_session_token_response_parses_nested_data_correctly():
